+++ src/app/locales/en/components/file-browser/item-features.js
@@ -1,0 +1,16 @@
+export default {
+  fileType: {
+    file: 'file',
+    dir: 'directory',
+    symlink: 'symbolic link',
+  },
+  status: {
+    qos: 'QoS',
+    dataset: 'Dataset',
+  },
+  protectionFlagsInfo: {
+    metadata: 'This {{fileType}}\'s metadata is write protected.',
+    data: 'This {{fileType}}\'s data is write protected.',
+    both: 'This {{fileType}}\'s data and metadata are write protected.',
+  },
+};