+++ src/app/locales/en/components/filesystem-browser.js
@@ -1,10 +1,12 @@
 import tableRowStatusBar from './filesystem-browser/table-row-status-bar';
+import tableHeadStatusBar from './filesystem-browser/table-head-status-bar';
 import tableHeadRow from './filesystem-browser/table-head-row';
 import emptyDir from './filesystem-browser/empty-dir';
 import tableRowColumns from './filesystem-browser/table-row-columns';
 
 export default {
   tableRowStatusBar,
+  tableHeadStatusBar,
   tableHeadRow,
   emptyDir,
   tableRowColumns,