+++ src/app/locales/en/components/filesystem-browser/table-head-status-bar.js
@@ -1,0 +1,3 @@
+// FIXME: maybe to remove
+
+export default {};